+++ packages/camera/camera_android_camerax/CHANGELOG.md
@@ -1,3 +1,5 @@
 ## NEXT
 
 * Creates camera_android_camerax plugin for development.
+* Adds CameraInfo class and removes unnecessary code from plugin.
+* Adds CameraSelector class.